Add insertion sort subroutine to bucket sort
