+++ src/visualizer/sortingvisualizer/util/SwappingAlgoUtil.js
@@ -2,7 +2,7 @@
 import { resetArray } from './GeneralUtil';
 import { swap } from '../../algorithm/sortingalgorithms/swap';
 
-const highlight = (firstIdx, secondIdx, arr) => {
+export const highlight = (firstIdx, secondIdx, arr) => {
   arr[firstIdx].isSwap = true;
   arr[secondIdx].isSwap = true;
 };
@@ -15,4 +15,24 @@ export const executeSwap = (firstIdx, secondIdx, arr, isSwapOccurring, algo) => 
   }
   swap(firstIdx, secondIdx, newTempArr);
   return newTempArr;
+};
+
+export const executeGenericSort = (
+  currentAnimation,
+  referenceArray,
+  visualizerAlgorithm,
+  setReferenceArray,
+) => {
+  let firstIdx = currentAnimation[0];
+  let secondIdx = currentAnimation[1];
+  let isSwapOccurring = currentAnimation[2];
+  let arrToUse = executeSwap(
+    firstIdx,
+    secondIdx,
+    referenceArray,
+    isSwapOccurring,
+    visualizerAlgorithm,
+  );
+  setReferenceArray(arrToUse);
+  return arrToUse;
 };
